requirements.txt
Django==4.2.25
djangorestframework
requests
python-dotenv
groq
gunicorn
